Ensure report creator uses latest JSON output format
Added a check for each test that ensures that the report creator
 is using  the latest format of the JSON output  of the compare 
command, ensuring the report creator is up to date with the command.
To support this case a testsutils package was created including test code
 that was located in the compare package, and now moved to the new
 package to allow usage in the report creator.
