+++ addon-tools/report-creator/report/testdata/DiffTestSuiteCreationWhenThereAreDiffs
@@ -1,1 +1,1 @@
-{"Summary":{"RequiredCRS":{},"NumMissing":0,"UnmatchedCRS":[],"NumDiffCRs":1,"TotalCRs":1},"Diffs":[{"DiffOutput":"diff -u -N /tmp/MERGED-3042598835/v1_configmap_kubernetes-dashboard_kubernetes-dashboard-settings /tmp/LIVE-442912111/v1_configmap_kubernetes-dashboard_kubernetes-dashboard-settings\n--- /tmp/MERGED-3042598835/v1_configmap_kubernetes-dashboard_kubernetes-dashboard-settings\t2024-07-24 13:48:39.476294975 +0300\n+++ /tmp/LIVE-442912111/v1_configmap_kubernetes-dashboard_kubernetes-dashboard-settings\t2024-07-24 13:48:39.476294975 +0300\n@@ -2,6 +2,6 @@\n kind: ConfigMap\n metadata:\n   labels:\n-    k8s-app: kubernetes-dashboardfunction was called successfully from different file\n+    k8s-app: kubernetes-dashboard\n   name: kubernetes-dashboard-settings\n   namespace: kubernetes-dashboard\n","CorrelatedTemplate":"cm.yaml","CRName":"v1_ConfigMap_kubernetes-dashboard_kubernetes-dashboard-settings"}]}
+{"Summary":{"RequiredCRS":{},"NumMissing":0,"UnmatchedCRS":[],"NumDiffCRs":1,"TotalCRs":1},"Diffs":[{"DiffOutput":"diff -u -N TEMP/v1_configmap_kubernetes-dashboard_kubernetes-dashboard-settings TEMP/v1_configmap_kubernetes-dashboard_kubernetes-dashboard-settings\n--- TEMP/v1_configmap_kubernetes-dashboard_kubernetes-dashboard-settings\tDATE\n+++ TEMP/v1_configmap_kubernetes-dashboard_kubernetes-dashboard-settings\tDATE\n@@ -2,6 +2,6 @@\n kind: ConfigMap\n metadata:\n   labels:\n-    k8s-app: kubernetes-dashboardfunction was called successfully from different file\n+    k8s-app: kubernetes-dashboard\n   name: kubernetes-dashboard-settings\n   namespace: kubernetes-dashboard\n","CorrelatedTemplate":"cm.yaml","CRName":"v1_ConfigMap_kubernetes-dashboard_kubernetes-dashboard-settings"}]}
