+++ addon-tools/report-creator/report/testdata/DiffTestSuiteCreationWhenThereAreDiffs.golden
@@ -4,7 +4,7 @@
 		<properties></properties>
 		<testcase classname="Matching Reference CR: cm.yaml" name="CR: v1_ConfigMap_kubernetes-dashboard_kubernetes-dashboard-settings" TIME>
 			<properties></properties>
-			<failure message="Differences found in CR: v1_ConfigMap_kubernetes-dashboard_kubernetes-dashboard-settings, Compared To Reference CR: cm.yaml" type="Difference">diff -u -N /tmp/MERGED-3042598835/v1_configmap_kubernetes-dashboard_kubernetes-dashboard-settings /tmp/LIVE-442912111/v1_configmap_kubernetes-dashboard_kubernetes-dashboard-settings&#xA;--- /tmp/MERGED-3042598835/v1_configmap_kubernetes-dashboard_kubernetes-dashboard-settings&#x9;2024-07-24 13:48:39.476294975 +0300&#xA;+++ /tmp/LIVE-442912111/v1_configmap_kubernetes-dashboard_kubernetes-dashboard-settings&#x9;2024-07-24 13:48:39.476294975 +0300&#xA;@@ -2,6 +2,6 @@&#xA; kind: ConfigMap&#xA; metadata:&#xA;   labels:&#xA;-    k8s-app: kubernetes-dashboardfunction was called successfully from different file&#xA;+    k8s-app: kubernetes-dashboard&#xA;   name: kubernetes-dashboard-settings&#xA;   namespace: kubernetes-dashboard&#xA;</failure>
+			<failure message="Differences found in CR: v1_ConfigMap_kubernetes-dashboard_kubernetes-dashboard-settings, Compared To Reference CR: cm.yaml" type="Difference">diff -u -N TEMP/v1_configmap_kubernetes-dashboard_kubernetes-dashboard-settings TEMP/v1_configmap_kubernetes-dashboard_kubernetes-dashboard-settings&#xA;--- TEMP/v1_configmap_kubernetes-dashboard_kubernetes-dashboard-settings&#x9;DATE&#xA;+++ TEMP/v1_configmap_kubernetes-dashboard_kubernetes-dashboard-settings&#x9;DATE&#xA;@@ -2,6 +2,6 @@&#xA; kind: ConfigMap&#xA; metadata:&#xA;   labels:&#xA;-    k8s-app: kubernetes-dashboardfunction was called successfully from different file&#xA;+    k8s-app: kubernetes-dashboard&#xA;   name: kubernetes-dashboard-settings&#xA;   namespace: kubernetes-dashboard&#xA;</failure>
 		</testcase>
 	</testsuite>
 	<testsuite tests="1" failures="0" TIME name="Missing Cluster Resources" TIME>
